requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
Flask==3.0.3
google-api-core==2.25.1
google-api-python-client==2.127.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-secret-manager==2.24.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.0.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
